pyproject.toml
[tool.poetry]
name = "qnabot"
version = "0.1.0"
description = ""
authors = ["Irfan Shah <irfan@pinelabs.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
python-environ = "^0.4.54"
pyautogen = {extras = ["retrievechat"], version = "^0.2.33"}
unstructured = {extras = ["all-docs"], version = "^0.15.0"}
chromadb = "^0.5.5"
litellm = {extras = ["proxy"], version = "^1.42.5"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
